+++ src/packages/griditem/griditem.tsx
@@ -1,41 +1,20 @@
-import React, {
-  CSSProperties,
-  FunctionComponent,
-  ReactNode,
-  useContext,
-} from 'react'
+import React, { CSSProperties, FunctionComponent, useContext } from 'react'
 import classNames from 'classnames'
-import { useConfig } from '@/packages/configprovider'
-import GridContext from '../grid/grid.context'
-import { BasicComponent } from '@/utils/typings'
+import GridContext from '../grid/context'
 import { pxCheck } from '@/utils/px-check'
-
-type GridDirection = 'horizontal' | 'vertical'
-
-export interface GridItemProps extends BasicComponent {
-  text: string | ReactNode
-  index: number
-  columns: string | number
-  gap: string | number
-  center: boolean
-  square: boolean
-  reverse: boolean
-  direction: GridDirection
-}
+import { GridItemProps } from './types'
 
 const defaultProps = {
   text: '',
   columns: 4,
   gap: 0,
   center: true,
-  square: false,
-  reverse: false,
   direction: 'vertical',
 } as GridItemProps
 export const GridItem: FunctionComponent<
   Partial<GridItemProps> & React.HTMLAttributes<HTMLDivElement>
 > = (props) => {
-  const { locale } = useConfig()
+  const classPrefix = 'nut-grid-item'
   const {
     children,
     style,
@@ -54,8 +33,6 @@ export const GridItem: FunctionComponent<
     ...defaultProps,
     ...props,
   }
-  const classPrefix = 'nut-grid-item'
-  const classes = classNames(classPrefix, className)
   const context = useContext(GridContext)
 
   const rootStyle = () => {
@@ -77,14 +54,17 @@ export const GridItem: FunctionComponent<
   }
 
   const contentClass = () => {
-    return classNames(`${classPrefix}-content`, {
-      [`${classPrefix}-content-border`]: true,
-      [`${classPrefix}-content-surround`]: gap,
-      [`${classPrefix}-content-center`]: center,
-      [`${classPrefix}-content-square`]: square,
-      [`${classPrefix}-content-reverse`]: reverse,
-      [`${classPrefix}-content-${direction}`]: !!direction,
-    })
+    return classNames(
+      `${classPrefix}-content`,
+      `${classPrefix}-content-border`,
+      {
+        [`${classPrefix}-content-surround`]: gap,
+        [`${classPrefix}-content-center`]: center,
+        [`${classPrefix}-content-square`]: square,
+        [`${classPrefix}-content-reverse`]: reverse,
+        [`${classPrefix}-content-${direction}`]: !!direction,
+      }
+    )
   }
 
   const handleClick = (e: any) => {
@@ -107,13 +87,13 @@ export const GridItem: FunctionComponent<
 
   return (
     <div
-      className={classes}
+      className={classNames(classPrefix, className)}
       style={rootStyle()}
-      {...rest}
       onClick={handleClick}
+      {...rest}
     >
       <div className={contentClass()}>
-        {children && <>{children}</>}
+        {children}
         {text && <div className={`${classPrefix}-text`}>{text}</div>}
       </div>
     </div>
