+++ src/packages/griditem/index.taro.ts
@@ -1,4 +1,4 @@
 import { GridItem } from './griditem.taro'
 
-export type { GridItemProps } from './griditem.taro'
+export type { GridItemProps } from './types'
 export default GridItem
