+++ src/packages/griditem/index.ts
@@ -1,4 +1,4 @@
 import { GridItem } from './griditem'
 
-export type { GridItemProps } from './griditem'
+export type { GridItemProps } from './types'
 export default GridItem
